requirements.txt
Cheetah==2.4.4
Flask==0.11.1
Flask-Ask==0.8.2
Jinja2==2.8
Landscape-Client==14.12
Markdown==2.6.7
MarkupSafe==0.23
PAM==0.4.2
PyYAML==3.10
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
Werkzeug==0.11.11
aniso8601==1.2.0
apt-xapian-index==0.45
argparse==1.2.1
awscli==1.11.28
beautifulsoup4==4.5.1
boto3==1.4.2
botocore==1.4.85
bs4==0.0.1
chardet==2.0.1
click==6.6
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
cssselect==1.0.0
docutils==0.13.1
funcsigs==1.0.2
futures==3.0.5
html5lib==0.999
itsdangerous==0.24
jmespath==0.9.0
jsonpatch==1.3
jsonpointer==1.0
logilab-common==0.61.0
lxml==3.7.0
mock==2.0.0
multi-key-dict==2.0.3
oauth==1.0.1
pbr==1.10.0
phue==0.9
prettytable==0.7.2
py==1.4.20
pyOpenSSL==0.13
pyasn1==0.1.9
pycurl==7.19.3
pyserial==2.6
python-apt===0.9.3.5ubuntu2
python-dateutil==2.6.0
python-debian===0.1.21-nmu2ubuntu2
python-jenkins==0.4.13
requests==2.12.4
rsa==3.4.2
s3transfer==0.1.9
six==1.10.0
ssh-import-id==3.21
tox==1.6.0
twill==1.8.0
urllib3==1.7.1
virtualenv==1.11.4
wheel==0.24.0
wsgiref==0.1.2
xmltodict==0.10.2
zope.interface==4.0.5
